ingestion-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
aio-pika>=9.4.0
minio>=7.2.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
tiktoken>=0.8.0
openai>=1.58.0
